python/multiplatform/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dagger-io==0.2.0
exceptiongroup==1.0.1
frozenlist==1.3.3
gql==3.4.0
graphql-core==3.2.3
idna==3.4
multidict==6.0.2
Pygments==2.13.0
python-dateutil==2.8.2
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
strawberry-graphql==0.141.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
